+++ CHANGELOG.rst
@@ -23,7 +23,7 @@ Changed
 * Corrupt cookie logging:
 
   * Now independent of other cookie logging; no longer needs to meet cookie size threshold or sampling rate.
-  * New setting ``UNUSUAL_COOKIE_SAMPLING_LOG_CHUNK`` helps avoid truncated (non-decryptable) messages by splitting the output across multiple log messages.
+  * New setting ``UNUSUAL_COOKIE_HEADER_LOG_CHUNK`` helps avoid truncated (non-decryptable) messages by splitting the output across multiple log messages.
 
 [4.8.1] - 2022-05-06
 --------------------